refactor: move hooks into their own files
