+++ app/javascript/utility/hooks/useDocumentTitle.js
@@ -1,0 +1,14 @@
+/**
+ * @flow
+ */
+
+import { useEffect } from 'react'
+
+export default function useDocumentTitle (title: string) {
+  useEffect(
+    () => {
+      document.title = title
+    },
+    [title]
+  )
+}
