+++ app/javascript/utility/hooks/useToggle.js
@@ -1,0 +1,15 @@
+/**
+ * @flow
+ */
+
+import { useState } from 'react'
+
+export default function useToggle (initial: boolean = false) {
+  const [state, set] = useState(initial)
+
+  function toggle () {
+    set(state => !state)
+  }
+
+  return [state, toggle]
+}
